+++ node/src/lib.rs
@@ -1,6 +1,5 @@
 //! A fresh FRAME-based Substrate node, ready for hacking.
 
-pub mod chain_init;
 pub mod chain_spec;
 mod inherent_data;
 mod main_chain_follower;