+++ changelog.md
@@ -7,12 +7,20 @@ This changelog is based on [Keep A Changelog](https://keepachangelog.com/en/1.1.
 ## Changed
 
 * moved out some cli related code from `node` crate, in order to require less copy-paste in users nodes
+* removed USE_CHAIN_INIT code. Migration strategy is to remove copy-pasted and adapted code. It will not compile with vanilla polkadot-sdk, that we plan to use in future.
+* moved `sc-consensus-aura` from input-output-hk/polkadot-sdk fork to this repository,
+  to `sc-partner-chains-consensus-aura` and `sp-partner-chains-consensus-aura`.
+  This change requires migration of the node, PartnerChainsProposerFactory has to be used.
+  See `service.rs` in `partner-chains-node` crate for an example.
 
 ## Removed
 
 ## Fixed
 
 ## Added
+* ETCM-7811 - native token movement observability components: `sp-native-token-management` and
+`pallet-native-token-management` crates; data sources behind the `native-token` feature in
+`main-chain-follower-api` and `db-sync-follower` crates.
 
 # v1.0.0rc1